src/semantic-kernel-azure-agent/requirements.txt
# Add your project dependencies below
# Example:
# python-dotenv
# azure-ai
semantic-kernel
azure-identity